rename: `router` 역할을 하는 `Router.tsx` 파일 `App.tsx` 파일에서 분리 (#160)
